requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
PyJWT>=2.6.0
bcrypt>=4.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0
